Contain the emulator
