+++ src/main.rs
@@ -2,29 +2,13 @@
 #![allow(unsafe_code)]
 #![deny(clippy::all)]
 
-use std::sync::mpsc::channel;
 use std::sync::Arc;
-use std::time::Duration;
 
-use crate::settings::gui::ToGuiEvent;
-use crate::{input::gamepad::ToGamepadEvent, settings::gui::GuiEvent};
-
-use anyhow::Result;
-use audio::{Audio, AudioSender};
 use bundle::Bundle;
 
-use fps::RateCounter;
-
-use gui::MainGui;
-
-use input::sdl2_impl::Sdl2Gamepads;
-use input::Inputs;
 use nes_state::emulator::Emulator;
 
-use nes_state::FrameData;
-use sdl2::EventPump;
-use window::egui_winit_wgpu::Renderer;
-use window::{create_window, NESFrame, Size};
+use window::{create_window, Size};
 use winit::event::{Event, WindowEvent};
 use winit::event_loop::EventLoop;
 
@@ -68,29 +52,6 @@ async fn main() {
     std::process::exit(0);
 }
 
-fn init(renderer: &mut Renderer) -> Result<(MainGui, EventPump, AudioSender)> {
-    // Needed because: https://github.com/libsdl-org/SDL/issues/5380#issuecomment-1071626081
-    sdl2::hint::set("SDL_JOYSTICK_THREAD", "1");
-    // TODO: Perhaps do this to fix this issue: https://github.com/libsdl-org/SDL/issues/7896#issuecomment-1616700934
-    //sdl2::hint::set("SDL_JOYSTICK_RAWINPUT", "0");
-
-    let sdl_context = sdl2::init().map_err(anyhow::Error::msg)?;
-    let sdl_event_pump = sdl_context.event_pump().map_err(anyhow::Error::msg)?;
-
-    //TODO: Figure out a resonable latency
-    let mut audio = Audio::new(&sdl_context, Duration::from_millis(40), 44100)?;
-
-    let inputs = Inputs::new(Sdl2Gamepads::new(
-        sdl_context.game_controller().map_err(anyhow::Error::msg)?,
-    ));
-
-    let emulator = Emulator::start()?;
-    let audio_tx = audio.stream.start()?;
-
-    let main_gui = MainGui::new(renderer, emulator, inputs, audio);
-    Ok((main_gui, sdl_event_pump, audio_tx))
-}
-
 async fn run() -> anyhow::Result<()> {
     let event_loop = EventLoop::new()?;
     let window = Arc::new(create_window(
@@ -102,80 +63,7 @@ async fn run() -> anyhow::Result<()> {
         Size::new(NES_WIDTH_4_3 as f64, NES_HEIGHT as f64),
         &event_loop,
     )?);
-    let mut renderer = Renderer::new(window.clone()).await.expect("TED");
-    let (event_tx, event_rx) = channel();
-    let _ = std::thread::Builder::new()
-        .name("Emulator".into())
-        .spawn(move || {
-            let (mut main_gui, mut sdl_event_pump, audio_tx) = init(&mut renderer).expect("TODO");
-            let mut nes_frame = NESFrame::new();
-            let mut rate_counter = RateCounter::new();
-
-            loop {
-                rate_counter.tick("Loop");
-                puffin::GlobalProfiler::lock().new_frame();
-                #[cfg(feature = "debug")]
-                puffin::profile_function!("Render");
-
-                for sdl_gui_event in sdl_event_pump
-                    .poll_iter()
-                    .flat_map(|e| e.to_gamepad_event())
-                    .map(GuiEvent::Gamepad)
-                {
-                    main_gui.handle_event(&sdl_gui_event, &renderer.window);
-                }
-
-                for winit_window_event in event_rx.try_iter() {
-                    match &winit_window_event {
-                        WindowEvent::Resized(physical_size) => {
-                            renderer.resize(*physical_size);
-                        }
-                        winit_window_event => {
-                            if !renderer
-                                .egui
-                                .handle_input(&renderer.window, winit_window_event)
-                                .consumed
-                            {
-                                if let Some(winit_gui_event) = winit_window_event.to_gui_event() {
-                                    main_gui.handle_event(&winit_gui_event, &renderer.window);
-                                }
-                            }
-                        }
-                    }
-                }
-
-                use crate::nes_state::NesStateHandler;
-
-                let joypads = &main_gui.inputs.joypads;
-                {
-                    #[cfg(feature = "debug")]
-                    puffin::profile_scope!("advance");
-                    let mut frame_data = main_gui
-                        .emulator
-                        .nes_state
-                        .advance(*joypads, &mut Some(&mut nes_frame));
-                    {
-                        #[cfg(feature = "debug")]
-                        puffin::profile_scope!("push audio");
-                        if let Some(FrameData { audio }) = &mut frame_data {
-                            log::trace!("Pushing {:} audio samples", audio.len());
-                            for s in audio {
-                                let _ = audio_tx.send(*s);
-                            }
-                        }
-                    }
-                }
-                {
-                    rate_counter.tick("Render");
-                    #[cfg(feature = "debug")]
-                    puffin::profile_scope!("render");
-                    main_gui.render_gui(&mut renderer, &nes_frame);
-                }
-                if let Some(report) = rate_counter.report() {
-                    println!("{report}");
-                }
-            }
-        });
+    let event_tx = Emulator::start(window.clone()).await?;
 
     Ok(event_loop.run(|winit_event, control_flow| {
         if let Event::WindowEvent {
@@ -192,7 +80,9 @@ async fn run() -> anyhow::Result<()> {
                     #[cfg(windows)]
                     window.request_redraw();
                 }
-                window_event => event_tx.send(window_event.clone()).expect("TODO"),
+                window_event => {
+                    let _ = event_tx.send(window_event.clone());
+                }
             }
         };
     })?)
